[sift] Forego emulation when there is no SIFT pipe
This should only happen after ROI end, so we don't care about what the benchmark does anyway.
A cleaner solution might be to call exit(0) to stop the benchmark entirely at the end of endROI(),
but this can freak out the MPI runtime, taking other (potentially still in-ROI) MPI ranks down with us.