[tools] Do not hardcode absolute paths to dependencies
tools/checkdependencies.py used to have some possible paths to
dependencies hardcoded, but this approach is inherently limited as the
user may install dependencies into arbitrary paths. Sometimes it is not
possible to install dependencies to a well-known path because of file
permissions.
Instead, let the platform resolve paths for dependencies as other build
systems typically do.