[rob] Fix CPI component decision
Even when all of the uops in the ROB are stalled due to memory
components, we have a chance to execute more uops if there is some room
that is greater than or equals to the dispatch width in the RS and ROB.
Therefore, the latency should be attribtued to memory components only
when there is no such room in the RS and ROB.

However, the check for the RS/ROB space was not performed when the
frontend stalls, and frontend stalls were incorrectly reported as memory
access stalls since the stalls caused by memory components take
precedence over the fronrend stalls.

For example, consider a loop with loads. In such a case, RobTimer only
attributes the latency to a memory component only when a load is stuck
and makes the ROB full.

If the loop contains frequent branch mispredictions, the ROB starves for
uops, and the latency caused by this starvation needs to be attributed
to the branch mispredictions. However, if the head of the ROB happens
to be a load uop, the latency will be incorrectly attributed to a
memory component. Since it is more likely that the ROB starves instead
it becomes full due to a stuck load uop, less latency should be
attributed to memory components, but RobTimer does otherwise.

The reported CPI stacks will also be useless for optmization; it looks
like we need to speed the load up, but it will only make the ROB starve
early. In contrary, eliminating the mispedicted branches by
if-conversion will actually resolve the ROB starvation and improve the
performance.

Fix this by always checking the RS/ROB space before attributing the
latency to memory components.

This commit also adds a check to tell if the mispredicted branch is kept
unresolved. If it is, it is likely that the branch resolution depends
on the stuck memory access so attribute the latency to it whether
the RS and the ROB have some room.