[rob] Remove cpiSerialization
The latency was used to be attributed to cpiSerialization when a
serialization or memory barrier instruction is the first uop in the
ROB. However, when such an instruction reaches to the head of the ROB,
it does no longer have the power to stop issuing and the potential
latency caused by the serialization has already been attributed to
preceeding load or store uops. As such, cpiSerialization gives no
insight for the impact of serialization or memory barrier.